refactor: move eservice notification handler to dedicated directory
